requirements.txt
requests>=2.31.0
sentence-transformers>=2.7.0
numpy>=1.26.0
python-dotenv>=1.0.0
openai>=1.0.0
msgpack>=1.0.0
pymupdf>=1.23.0
python-docx>=1.1.0
pytest>=7.0.0
